requirements.txt
hypercorn
fastapi
tortoise-orm
pydantic_settings
python-decouple
email-validator
